update at folder system

+++ S09/GeekShopping/GeekShopping.ProductAPI/Model/Base/BaseEntity.cs
@@ -5,9 +5,9 @@ namespace GeekShopping.ProductAPI.Model.Base
 {
     public class BaseEntity
     {
+
         [Key]
         [Column("id")]
-        public int Id { get; set; }
-
+        public long Id { get; set; }
     }
 }
